fix(lint): resolve all golangci-lint errors
- Fixed field alignment in Fetcher and FetchResult structs
- Renamed unused parameter in CheckRedirect to _
- Use http.NoBody instead of nil for GET requests
- Added proper error handling for resp.Body.Close()
- Fixed integer overflow conversions using explicit int32 cast
- Fixed gofmt formatting in placeholder_test.go

All tests passing locally.

+++ internal/placeholder/placeholder.go
@@ -124,8 +124,8 @@ func addLabel(img *image.RGBA, label string, col color.Color) {
 	}
 
 	point := fixed.Point26_6{
-		X: fixed.Int26_6(x * 64),
-		Y: fixed.Int26_6(y * 64),
+		X: fixed.Int26_6(int32(x) * 64),
+		Y: fixed.Int26_6(int32(y) * 64),
 	}
 
 	d := &font.Drawer{
